+++ cfbd_json_py/rankings.py
@@ -1,6 +1,6 @@
 
 def get_cfbd_poll_rankings(
-        year: int,
+        season: int,
         api_key: str = None,
         api_key_dir: str = None,
         week: int = None,