+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # CHANGELOG: cfbd_json_py
 
+## 0.0.6: The "Drives" Update.
+
+- Implemented `cfbd_json_py.drives.get_cfbd_drives_info()`, a function that allows a user to get CFB drive data from the CFBD API.
+- Added a sample python file for `cfbd_json_py.draft.get_cfbd_nfl_draft_info()` in the "Usage" section for the function's description, as well as cleaning up the code formatting for this function.
+- Fixed a minor spelling error in the function description for `cfbd_json_py.betting.get_cfbd_betting_lines()`
+- Updated the package version to `0.0.6`.
+
 ## 0.0.5: The "Draft" Update.
 
 - Implemented `cfbd_json_py.draft.get_cfbd_nfl_teams()`, a function that allows a user to get a list of NFL teams from the CFBD API.
 - Implemented `cfbd_json_py.draft.get_cfbd_nfl_positions()`, a function that allows a user to get a list of player positions for the NFL Draft from the CFBD API.
-- Implemented `cfbd_json_py.draft.get_cfbd_nfl_positions()`, a function that allows a user to get a list of actual NFL Draft selections from the CFBD API.
+- Implemented `cfbd_json_py.draft.get_cfbd_nfl_draft_info()`, a function that allows a user to get a list of actual NFL Draft selections from the CFBD API.
 - Updated the function descriptions for the following functions:
   - `cfbd_json_py.betting.get_cfbd_betting_lines()`
   - `cfbd_json_py.coaches.get_cfbd_coaches_info()`